Move out queries folder and add _public tag to top level pages
