+++ app/components/FFXIVResults/item-history/PriceHistoryChart.tsx
