+++ app/components/FFXIVResults/item-history/SaleHistoryTable.tsx
