+++ app/components/FFXIVResults/listings/Differences.tsx
