+++ app/components/form/select/SelectWorld/index.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react'
 import * as locations from '~/utils/locations'
 import type { Transition } from '@remix-run/react/dist/transition'
 import type { ValidationResult } from 'remix-validated-form'
-import type SelectWorldInputFields from '~/routes/options'
+import type SelectWorldInputFields from '~/routes/_public.options'
 import type { GetDeepProp } from '~/utils/ts'
 import type { WorldsList } from '~/utils/locations/Worlds'
 import { SelectDataCenter } from '~/components/form/select/SelectWorld/SelectDataCenter'
