+++ app/routes/queries.item-data.$itemId.tsx
@@ -16,9 +16,9 @@ import type { ListingResponseType } from '~/requests/GetListing'
 import GetListing from '~/requests/GetListing'
 import { getUserSessionData } from '~/sessions'
 import { getItemNameById } from '~/utils/items'
-import HistoryResults from '../item-history/Results'
+import HistoryResults from '~/components/FFXIVResults/item-history/Results'
 import NoResults from '~/components/Common/NoResults'
-import ListingResults from '../listings/Results'
+import ListingResults from '~/components/FFXIVResults/listings/Results'
 import { ErrorBoundary as ErrorBounds } from '~/components/utilities/ErrorBoundary'
 import { useTypedSelector } from '~/redux/useTypedSelector'
