+++ app/routes/queries.item-history.tsx
@@ -12,7 +12,7 @@ import ItemSelect from '~/components/Common/ItemSelect'
 import type { ItemSelected } from '~/components/Common/ItemSelect'
 import { useEffect, useState } from 'react'
 import SmallFormContainer from '~/components/form/SmallFormContainer'
-import Results from './Results'
+import Results from '~/components/FFXIVResults/item-history/Results'
 import { PageWrapper, Title } from '~/components/Common'
 import { useDispatch } from 'react-redux'
 import { setItemHistory } from '~/redux/reducers/queriesSlice'
