+++ app/routes/queries.world-comparison.tsx
