--- app/routes/queries/item-history.tsx
@@ -1,9 +1,0 @@
-import { Outlet } from '@remix-run/react'
-
-export default function FullScan() {
-  return (
-    <div>
-      <Outlet />
-    </div>
-  )
-}
